I have passed all of the falcor tests except for the fact I have no progressively code.  This is alarming :)
