+++ lib/request/sendSetRequest.js
@@ -59,7 +59,7 @@ var sendSetRequest = function(originalJsonGraph, model, callback) {
                 };
             }), null, errorSelector, comparator);
 
-            model._path = bound;
+            model._path = boundPath;
 
             callback(dataSourceError);
         }, function onCompleted() {
