+++ lib/response/set/SetResponse.js
@@ -86,26 +86,12 @@ SetResponse.prototype = Object.create(Observable.prototype);
 SetResponse.prototype._subscribe = function _subscribe(observer) {
     var groups = this._groups;
     var model = this._model;
-    var requestedAndOptimizedPaths = setGroupsIntoCache(model, groups);
-    var optimizedPaths = requestedAndOptimizedPaths.optimizedPaths;
-    var requestedPaths = requestedAndOptimizedPaths.requestedPaths;
     var isJSONGraph = this._isJSONGraph;
     var isProgressive = this._isProgressive;
 
-    // Local set only.  We perform a follow up get.  If performance is ever
-    // a requirement simply requiring in checkCacheAndReport and use get request
-    // internals.  Figured this is more 'pure'.
-    if (this._isMaster) {
-        var get = new GetResponse(model, requestedPaths,
-                                  isJSONGraph, isProgressive);
-
-        return get.subscribe(observer);
-    }
-
     // Starts the async request cycle.
     return setRequestCycle(
-        model, observer, requestedPaths, optimizedPaths,
-        isJSONGraph, isProgressive, 0);
+        model, observer, groups, isJSONGraph, isProgressive, 0);
 };
 
 /**
