+++ lib/response/set/setRequestCycle.js
@@ -1,18 +1,33 @@
 var emptyArray = [];
 var AssignableDisposable = require('./../AssignableDisposable');
 var GetResponse = require('./../get/GetResponse');
+var setGroupsIntoCache = require('./setGroupsIntoCache');
 
 /**
  * The request cycle for set.  This is responsible for requesting to dataSource
  * and allowing disposing inflight requests.
  */
-module.exports = function setRequestCycle(model, observer, requestedPaths,
-                                          optimizedPaths, isJSONGraph,
-                                          isProgressive, count) {
+module.exports = function setRequestCycle(model, observer, groups,
+                                          isJSONGraph, isProgressive, count) {
     // we have exceeded the maximum retry limit.
     if (count === 10) {
         throw new MaxRetryExceededError();
     }
+
+    var requestedAndOptimizedPaths = setGroupsIntoCache(model, groups);
+    var optimizedPaths = requestedAndOptimizedPaths.optimizedPaths;
+    var requestedPaths = requestedAndOptimizedPaths.requestedPaths;
+    var isMaster = model._source === undefined;
+
+    // Local set only.  We perform a follow up get.  If performance is ever
+    // a requirement simply requiring in checkCacheAndReport and use get request
+    // internals.  Figured this is more 'pure'.
+    if (isMaster) {
+        var get = new GetResponse(model, requestedPaths,
+                                  isJSONGraph, isProgressive);
+        return get.subscribe(observer);
+    }
+
 
     // TODO: Progressively here
     var currentJSONGraph = getJSONGraph(model, optimizedPaths);
@@ -46,6 +61,9 @@ module.exports = function setRequestCycle(model, observer, requestedPaths,
             // TODO: The unhappy case.  I am unsure how this can even be
             // achieved.
             else {
+                // We need to restart the setRequestCycle.
+                setRequestCycle(model, observer, groups, isJSONGraph,
+                                isProgressive, count + 1);
             }
         });
